Minor fixes, Added controller tests and api documentation.

+++ src/main/java/com/netcompany/accountmanagementsystem/model/Account.java
@@ -8,6 +8,6 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 @AllArgsConstructor
 public class Account {
-    private int accountId;
-    private int beneficiaryId;
+    private Long accountId;
+    private Long beneficiaryId;
 }
